+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.22.5dev
 * [Feature] Add binder badge in the comments
 
+* [Feature] Deprecates `ploomber cloud`
 * [Fix] Compatibility with mistune 3
 
 ## 0.22.4 (2023-06-01)